# server/swagger.yml
openapi: 3.0.0
info:
  title: Gamers Social Media Api
  version: '1.0'
servers:
  - url: /api
  - url: http://localhost:5000/api
components:
  schemas:
...
    room:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        creator:
          type: string
        game:
          type: string
        platform:
          type: string
        description:
          type: string
        members:
          type: number
        private:
          type: string
...
paths:
...
  /room/:
    post:
      summary: create a room
      tags:
      - rooms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                game:
                  type: string
                  description: room only accept games from the official games list
                platform:
                  type: string
                  description: only accept on of *case-sensitive* - Pc, Xbox, Playstation, Android, Apple, Psp
                description:
                  type: string
                private:
                  type: boolean
              required: 
                - name
                - game
                - platform
      responses:
        '200':
          description: room was created
          content:
            plain/text:
              schema:
                type: string
                description: the created room id
        '400':
          description: input validation
          content:
            text/plain:
              schema:
                type: string
              examples:
                formatting:
                  value: name, game, platform, description or private are not formatted correctly
        '404':
          description: user is not logged in
    get:
      summary: get a room by its id
      tags:
      - rooms
      parameters:
        - in: query
          description: room id
          name: roomId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: the room with the specified id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/room"
        '400':
          description: input error
          content:
            text/plain:
              schema:
                type: string
              examples:
                the specified room not found:
                  value: room not found
                room id is required:
                  value: no specified room id
    delete:
      summary: delete a room by its id
      tags:
      - rooms
      parameters:
        - in: query
          description: room id
          name: roomId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: the room was deleted
        '403':
          description: room dose not exist or user has no permission to delete room
          content:
            text/plain:
              schema:
                type: string
              examples:
                exist or permission:
                  value: room dose not exist or user have no permission to delete it
        '404':
          description: user is not logged in
    put:
      summary: edit a room by its id
      tags:
      - rooms
      parameters:
        - in: query
          description: room id
          name: roomId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                game:
                  type: string
                platform:
                  type: string
                description:
                  type: string
                private:
                  type: boolean
      responses:
        '200':
          description: saved changes
        '403':
          description: room dose not exist or user has no permission to edit room
          content:
            text/plain:
              schema:
                type: string
              examples:
                exist or permission:
                  value: room dose not exist or user have no permission to delete it
        '400':
          description: input error
          content:
            text/plain:
              schema:
                type: string
              examples:
                formatting:
                  value: name, game, platform, description or private are not formatted correctly
        '404':
          description: user is not logged in
...
  /search/rooms:
    get:
      summary: search for rooms
      tags:
      - Search
      parameters:
        - in: query
          description: room name or part of it
          name: name
          schema:
            type: string
          required: false
        - in: query
          description: room game
          name: game
          schema:
            type: string
          required: false
        - in: query
          description: room minimum participates number
          name: min_participates
          schema:
            type: number
          required: false
        - in: query
          description: room maximum participates number
          name: max_participates
          schema:
            type: number
          required: false
        - in: query
          description: room game platform
          name: platform
          schema:
            type: string
          required: false
        - in: query
          description: offset results by number
          name: skip
          schema:
            type: number
          required: false
        - in: query
          description: limit results number
          name: limit
          schema:
            type: number
          required: false
      responses:
        '200':
          description: filtered list of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/room'
...